Restricted delete operation to OWNER role. Performs additional check to make sure the user that you're attempting to delete actually exists in the database

+++ actions/delete-user.js
@@ -1,10 +1,26 @@
 "use server";
 
 import prisma from "@/lib/prismadb";
+import { currentRole } from "@/lib/auth";
+import { UserRole } from "@prisma/client";
 
 export const deleteUser = async (userId) => {
+  const role = await currentRole();
+
+  if (role !== UserRole.OWNER) {
+    return { error: "You don't have permission to delete a user!" };
+  }
+
   if (!userId) {
     return { error: "UserId not found!" }
+  }
+
+  const existingUser = await prisma.user.findUnique({
+    where: { id: userId }
+  });
+
+  if (!existingUser) {
+    return { error: "User not found!" };
   }
 
   await prisma.user.delete({
